requirements/base.txt
Django==2.2
djangorestframework==3.12.2
django-filter==21.1
